E2E Framework: Folders cleanup and shuffling
