+++ tests/end_to_end/test_suites/sample_tests.py
@@ -15,7 +15,7 @@ log = logging.getLogger(__name__)
 # ** IMPORTANT **: This is just an example on how to add a test with below pre-requisites.
 # Task Runner API Test function for federation run using sample_model
 # 1. Create OpenFL workspace, if not present for the model and add relevant dataset and its path in plan/data.yaml
-# 2. Append the model name to ModelName enum in tests/end_to_end/utils/constants.py
+# 2. Append the model name to ModelName enum in tests/end_to_end/utils/defaults.py
 # 3. a. Use fx_federation_tr fixture for task runner with bare metal or docker approach.
 # 3. b. Use fx_federation_tr_dws fixture for task runner with dockerized workspace approach.
 # 4. Fixture will contain - model_owner, aggregator, collaborators, workspace_path, local_bind_path
