+++ tests/end_to_end/utils/tr_common_fixtures.py
@@ -3,6 +3,7 @@
 
 import pytest
 
+import tests.end_to_end.utils.data_helper as data_helper
 from tests.end_to_end.utils.tr_workspace import create_tr_workspace, create_tr_dws_workspace
 
 
@@ -38,3 +39,30 @@ def fx_federation_tr_dws(request):
     """
     request.config.test_env = "task_runner_dockerized_ws"
     return create_tr_dws_workspace(request)
+
+
+@pytest.fixture(scope="function")
+def fx_verifiable_dataset_with_s3(request):
+    """
+    Fixture for verifiable dataset with S3 bucket.
+    Note: As this is a function level fixture, thus no import is required at test level.
+    """
+    return data_helper.prepare_verifiable_dataset(request, dataset_type="s3")
+
+
+@pytest.fixture(scope="function")
+def fx_verifiable_dataset_with_azure_blob(request):
+    """
+    Fixture for verifiable dataset with Azure Blob Storage.
+    Note: As this is a function level fixture, thus no import is required at test level.
+    """
+    return data_helper.prepare_verifiable_dataset(request, dataset_type="azure_blob")
+
+
+@pytest.fixture(scope="function")
+def fx_verifiable_dataset_with_all(request):
+    """
+    Fixture for verifiable dataset with all combinations of S3, Azure Blob Storage and local data.
+    Note: As this is a function level fixture, thus no import is required at test level.
+    """
+    return data_helper.prepare_verifiable_dataset(request, dataset_type="all")
